+++ src/lib/wallet/helpers.ts
